Alterei campo do preço na tabela dos produtos

+++ database/migrations/2018_01_04_085954_create_products_table.php
@@ -19,8 +19,8 @@ class CreateProductsTable extends Migration
             $table->string('slug')->nullable();
             $table->string('image', 200)->nullable();
             $table->string('details')->nullable();
-            $table->double('old_price', 10, 2);
-            $table->double('price', 10, 2);
+            $table->decimal('old_price',10,2);
+            $table->decimal('price', 10, 2);
             $table->text('description')->nullable();
             $table->boolean('featured')->default(false);
             $table->boolean('active')->default(true); //s ou n
